requirements.txt
absl-py==0.11.0
astunparse==1.6.3
bokeh==2.2.3
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
chromedriver-install==1.0.3
click==7.1.2
cycler==0.10.0
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
flake8==3.8.4
Flask==1.1.2
Flask-Compress==1.8.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
nibabel==3.2.1
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
Pillow==8.0.1
plotly==4.11.0
protobuf==3.14.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycodestyle-magic==0.5
pyderman==2.0.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
scipy==1.5.3
seaborn==0.11.0
selenium==3.141.0
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
